requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
openai==1.70.0
psycopg2-binary==2.9.10
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.24.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
Werkzeug==3.1.3
